Development

+++ controller/user.controller.js
@@ -77,17 +77,18 @@ exports.createUser = async function (req, res, next) {
 
 exports.deleteUser = function (req, res, next) {
   try {
-    const {email} = req.params;
+    const currentUser = req.body.email;
     if (req.user.role === 'manager') {
-      UserModel.findOneAndDelete(email).then(data => {
+      console.log(currentUser);
+      UserModel.findOneAndDelete({email: currentUser}).then(data => {
         if (!data) {
           return res
             .status(400)
-            .send({message: `User with email=${email} was not found`});
+            .send({message: `User with email=${currentUser} was not found`});
         }
-        res
-          .status(200)
-          .send({message: `User with email=${email} was deleted successfully`});
+        res.status(200).send({
+          message: `User with email=${currentUser} was deleted successfully`,
+        });
       });
     }
   } catch (error) {
@@ -120,11 +121,13 @@ exports.getAllUsers = function (req, res, next) {
 // Manager can update user information (not id, password)
 // Gardeners can update user information (not id, email, password)
 
-exports.updateUser = (req, res, next) => {
-  const {email} = req.body;
+exports.updateUser = async (req, res, next) => {
+  const userEmail = req.body.selectedUser;
   const {password} = req.body;
+  const currentUser = req.user.email;
+  console.log(req.body);
   try {
-    if (req.user.role === 'manager') {
+    if (req.body.place === 'dashboard') {
       // update a user by its email
       if (!req.body)
         return res
@@ -135,29 +138,33 @@ exports.updateUser = (req, res, next) => {
         return res.status(400).send({message: 'Cannot update password'});
 
       // Update user based on email
-      UserModel.findOneAndUpdate(email, req.body, {
+      await UserModel.findOneAndUpdate({email: userEmail}, req.body, {
         useFindAndModify: false,
       }).then(data => {
         if (!data) {
           res.status(404).send({
-            message: `Cannot update user with email=${email}. Maybe the user was not found`,
+            message: `Cannot update user with email=${userEmail}. Maybe the user was not found`,
           });
+        } else {
+          res.status(200).send({data});
         }
       });
-    } else if (req.user.role === 'gardener') {
-      if (password || email)
+    } else {
+      if (currentUser)
         return res
           .status(400)
           .send({message: 'Cannot update password or email'});
 
       // Update user based on email
-      UserModel.findOneAndUpdate(email, req.body, {
+      await UserModel.findOneAndUpdate({email: currentUser}, req.body, {
         useFindAndModify: false,
       }).then(data => {
         if (!data) {
           res.status(404).send({
-            message: `Cannot update user with email=${email}. Maybe the user was not found`,
+            message: `Cannot update user with email=${currentUser}. Maybe the user was not found`,
           });
+        } else {
+          res.status(200).send({data});
         }
       });
     }
